Move DumpServiceTests to integration tests
