+++ Tests/TuistKitIntegrationTests/Commands/DumpServiceIntegrationTests.swift
@@ -8,20 +8,14 @@ import XCTest
 @testable import TuistSupportTesting
 
 final class DumpServiceTests: TuistTestCase {
-    var errorHandler: MockErrorHandler!
-    var subject: DumpService!
-    var manifestLoading: ManifestLoading!
+    private var subject: DumpService!
 
     override func setUp() {
         super.setUp()
-        errorHandler = MockErrorHandler()
-        manifestLoading = ManifestLoader()
-        subject = DumpService(manifestLoader: manifestLoading)
+        subject = DumpService()
     }
 
     override func tearDown() {
-        errorHandler = nil
-        manifestLoading = nil
         subject = nil
         super.tearDown()
     }
@@ -237,5 +231,78 @@ final class DumpServiceTests: TuistTestCase {
         """
 
         XCTAssertPrinterOutputContains(expected)
+    }
+
+    func test_run_throws_when_project_and_file_doesnt_exist() throws {
+        try assertLoadingRaisesWhenManifestNotFound(manifest: .project)
+    }
+
+    func test_run_throws_when_workspace_and_file_doesnt_exist() throws {
+        try assertLoadingRaisesWhenManifestNotFound(manifest: .workspace)
+    }
+
+    func test_run_throws_when_config_and_file_doesnt_exist() throws {
+        try assertLoadingRaisesWhenManifestNotFound(manifest: .config)
+    }
+
+    func test_run_throws_when_template_and_file_doesnt_exist() throws {
+        try assertLoadingRaisesWhenManifestNotFound(manifest: .template)
+    }
+
+    func test_run_throws_when_dependencies_and_file_doesnt_exist() throws {
+        try assertLoadingRaisesWhenManifestNotFound(manifest: .dependencies)
+    }
+
+    func test_run_throws_when_plugin_and_file_doesnt_exist() throws {
+        try assertLoadingRaisesWhenManifestNotFound(manifest: .plugin)
+    }
+
+    func test_run_throws_when_the_manifest_loading_fails() throws {
+        for manifest in DumpableManifest.allCases {
+            let tmpDir = try temporaryPath()
+            try "invalid config".write(
+                toFile: tmpDir.appending(component: manifest.manifest.fileName(tmpDir)).pathString,
+                atomically: true,
+                encoding: .utf8
+            )
+            XCTAssertThrowsError(try subject.run(path: tmpDir.pathString, manifest: manifest))
+        }
+    }
+
+    // MARK: - Helpers
+
+    private func assertLoadingRaisesWhenManifestNotFound(manifest: DumpableManifest) throws {
+        try fileHandler.inTemporaryDirectory { tmpDir in
+            var expectedDirectory = tmpDir
+            if manifest == .config {
+                expectedDirectory = expectedDirectory.appending(component: Constants.tuistDirectoryName)
+                if !fileHandler.exists(expectedDirectory) {
+                    try fileHandler.createFolder(expectedDirectory)
+                }
+            }
+            XCTAssertThrowsSpecific(
+                try subject.run(path: tmpDir.pathString, manifest: manifest),
+                ManifestLoaderError.manifestNotFound(manifest.manifest, expectedDirectory)
+            )
+        }
+    }
+}
+
+extension DumpableManifest {
+    var manifest: Manifest {
+        switch self {
+        case .project:
+            return .project
+        case .workspace:
+            return .workspace
+        case .config:
+            return .config
+        case .template:
+            return .template
+        case .dependencies:
+            return .dependencies
+        case .plugin:
+            return .plugin
+        }
     }
 }
